[rust-i] Runtime check, native clock and misc updates: 1.  Add runtime check to prevent reuse of const identifier in the same     scope temporarily as this will be moved to be a SyntaxError or Parsing     error in the future. Add new RuntimeError for this and temporary     in_current_scope method to Environment struct. 2.  Update native clock function to return milliseconds instead of seconds 3.  Change visibility of Environment struct fields as all access should     be through the getters and setters. 4.  More todos and docs

+++ rust/src/callables/native/clock/mod.rs
@@ -45,6 +45,6 @@ impl Callable for ClockFunc {
         _interpreter: &mut Interpreter,
         _arguments: Vec<Value>,
     ) -> Result<Value, RuntimeError> {
-        Ok(Value::Number(get_current_time() as f64))
+        Ok(Value::Number(get_current_time()))
     }
 }
